Feat Profile Fragment into User Info
프로필 변경 추가 추후 elastic ip 변경 후 수정예정
