+++ app/src/main/res/values/strings.xml
@@ -94,7 +94,6 @@
     <string name="onboarding_permission_scr_button_text">FITMATE 시작하기</string>
 
 
-
     <string name="my_fit_scr_title">My Fit</string>
     <string name="my_fit_scr_fit_off">피트 오프</string>
     <string name="my_fit_scr_fit_record">기록하기</string>
